alert colocado globalmente em routes
